src/image_ingestion_service/consumer_drivebc/requirements.txt
aio-pika
pillow
